+++ src/test/java/org/apiphany/ApiPredicatesTest.java
@@ -3,17 +3,14 @@ package org.apiphany;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import java.lang.reflect.InvocationTargetException;
 import java.util.Collections;
 import java.util.List;
 
+import org.apiphany.utils.Tests;
 import org.junit.jupiter.api.Test;
-import org.morphix.lang.JavaObjects;
 import org.morphix.reflection.Constructors;
-import org.morphix.reflection.ReflectionException;
 
 /**
  * Test class for {@link ApiPredicates}.
@@ -26,10 +23,7 @@ class ApiPredicatesTest {
 
 	@Test
 	void shouldThrowExceptionOnCallingConstructor() {
-		ReflectionException reflectionException =
-				assertThrows(ReflectionException.class, () -> Constructors.IgnoreAccess.newInstance(ApiPredicates.class));
-		InvocationTargetException invocationTargetException = JavaObjects.cast(reflectionException.getCause());
-		UnsupportedOperationException unsupportedOperationException = JavaObjects.cast(invocationTargetException.getCause());
+		UnsupportedOperationException unsupportedOperationException = Tests.verifyDefaultConstructorThrows(ApiPredicates.class);
 		assertThat(unsupportedOperationException.getMessage(), equalTo(Constructors.MESSAGE_THIS_CLASS_SHOULD_NOT_BE_INSTANTIATED));
 	}