+++ src/test/java/org/apiphany/security/oauth2/OAuth2TokenProviderConfigurationTest.java
@@ -1,0 +1,24 @@
+package org.apiphany.security.oauth2;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.equalTo;
+
+import org.apiphany.utils.Tests;
+import org.junit.jupiter.api.Test;
+import org.morphix.reflection.Constructors;
+
+/**
+ * Test class for {@link OAuth2TokenProviderConfiguration}.
+ *
+ * @author Radu Sebastian LAZIN
+ */
+class OAuth2TokenProviderConfigurationTest {
+
+	@Test
+	void shouldNotInstantiateDefaultClass() {
+		UnsupportedOperationException unsupportedOperationException =
+				Tests.verifyDefaultConstructorThrows(OAuth2TokenProviderConfiguration.Default.class);
+		assertThat(unsupportedOperationException.getMessage(), equalTo(Constructors.MESSAGE_THIS_CLASS_SHOULD_NOT_BE_INSTANTIATED));
+	}
+
+}